modif create sauce delete sauce

+++ backend/controllers/sauceControllers.js
@@ -15,26 +15,79 @@ exports.getOneSauce = (req, res) => {
 };
 
 exports.createSauce = (req, res) => {
+  const sauceObject = JSON.parse(req.body.modelSauceSchema);
+  delete sauceObject._id;
   const sauceSchema = new modelSauceSchema({
-    ...req.body,
+    ...sauceObject,
+    imageUrl: `${req.protocol}://${req.get("host")}/images/${
+      req.file.filename
+    }`,
   });
   sauceSchema
     .save()
-    .then(() => res.status(201).json({ message: "Object enregistré !" }))
+    .then(() => res.status(201).json({ message: "Sauce enregistré !" }))
     .catch((error) => res.status(400).json({ error }));
 };
 
 exports.updateOneSauce = (req, res) => {
+  // Faire un findOne comme sur le delete
+  //puis garder le nom de lancienne image dans une variable
   modelSauceSchema
-    .updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(200).json({ message: "Update sauce !" }))
+    .findOne({ _id: req.params.id })
+    .then((sauce) => {
+      console.log(sauce);
+      //si je n'ai pas trouvée de sauce je renvoie une 404
+      if (!sauce) {
+        return res.satus(404).send("sauce non trouvée");
+      }
+      const oldImg = sauce.imageUrl;
+      let sauceToUpdate;
+      if (req.file) {
+        sauceToUpdate = {
+          ...req.body,
+          imageUrl: `${req.protocol}://${req.get("host")}/images/${
+            req.file.filename
+          }`,
+        };
+      } else {
+        sauceToUpdate = { ...req.body };
+      }
+
+      modelSauceSchema
+        .updateOne(
+          { _id: req.params.id },
+          { ...sauceToUpdate, _id: req.params.id }
+        )
+        .then(() => {
+          if (req.file) {
+            //suppression de lancienne image : le nom de lancien image est stocké dans la variable oldImg
+          }
+          res.status(200).json({ message: "Update sauce !" });
+        })
+        .catch((error) => res.status(400).json({ error }));
+    })
     .catch((error) => res.status(400).json({ error }));
 };
 
 exports.deleteOneSauce = (req, res) => {
+  // avant de faire deleteOne faire un findOne
   modelSauceSchema
-    .deleteOne({ _id: req.params.id })
-    .then((modelSauceSchema) => res.status(200).json(modelSauceSchema))
+    .findOne({ _id: req.params.id })
+    .then((sauce) => {
+      console.log(sauce);
+      //si je n'ai pas trouvée de sauce je renvoie une 404
+      if (!sauce) {
+        return res.satus(404).send("sauce non trouvée");
+      }
+      //suppression de la sauce
+      modelSauceSchema
+        .deleteOne({ _id: req.params.id })
+        .then(() => {
+          // avant le res supprimer le fichier de la photo
+          res.status(200).json({ message: "Sauce supprimé !" });
+        })
+        .catch((error) => res.status(400).json({ error }));
+    })
     .catch((error) => res.status(400).json({ error }));
 };
 
@@ -44,6 +97,6 @@ exports.createOneLike = (req, res) => {
   });
   sauceSchema
     .save({ _id: req.params.id })
-    .then(() => res.status(201).json({ message: "Object enregistré !" }))
+    .then(() => res.status(201).json({ message: "Sauce enregistré !" }))
     .catch((error) => res.status(400).json({ error }));
 };
